requirements.txt
langchain-core>=1.0.0,<2.0.0
langchain>=1.0.0,<2.0.0
langgraph>=1.0.0,<2.0.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama
langchain-community
tavily-python
langchain-text-splitters>=1.1.0
pydantic>=2.0
pinecone>=5.4.2
pinecone-plugin-assistant>=3.0
pypdfium2>=5.0
python-docx>=1.0
pdfplumber>=0.11
presidio-analyzer>=2.2.360
presidio-anonymizer>=2.2.360
python-dotenv>=1.0
tiktoken>=0.8
nltk>=3.9
requests>=2.32
aiohttp>=3.10
